np.bool, np.int, np.float are deprecated
`np.bool`, `np.int`, and `np.float` are long-deprecated aliases for `bool` etc. Replace them to avoid deprecation warnings in NumPy 1.20+.

+++ runscripts/metrics/behavior_metrics/behavior_metrics.py
@@ -112,7 +112,7 @@ def find_indices_bulk(to_search, targets):
 	Returns:
 		The indices.
 	"""
-	return np.array([to_search.index(elem) for elem in targets], np.int)
+	return np.array([to_search.index(elem) for elem in targets], int)
 
 
 def np_pick(array, pick_spec):
